Fake the functions on the libclang API
I would like to be able to use test doubles for the translation unit
tests, but use the actual libclang API in end-to-end tests, so add the
ability to fake the libclang API at runtime. Note that the fakes will
not be compiled in release builds. So the tests that use fakes won't
work, but the faking has no runtime overhead.